Add files via upload
The code for the plot of Mental Illness During Covid as well as the plot encompass the key points of analysis of the model that looks at anxiety and depression from the CDC and KFF data bases on Mental Illness during the Covid 19 pandemic. The other discussion boxes in the repository have introductory, methods, results and conclusion comments to serve as a paper. But since this was an ongoing iterative process, I am reviewing these elements next to the key code and plot for the model entries. Intro:  Since the start of the Covid pandemic, enormous attention has  been focused on hospitalizations, respiratory arrest, and mortality outcomes. However, only casual attention has been given to morbidity and in particular mental illness. It is for this reason that this model was initiated. Methods: q 3 month percentage: anxiety,  depression, and anxiety and depression data from IHSS CDC questionnaires dating prior to June 2019 and thru Sept 2021 were extracted from CDC and KFF datasets and organized into a concise table for analysis.Results: As can be seen from the plot during the Covid 19 Pandemic over a 3 fold increase in anxiety and depression percentages were noted in the same populations compared to prior to the pandemic. Conclusion: Mental illness levels showed a dramatic increase during the Covid 19 pandemic.